@@ -20,14 +20,15 @@
         ]
     },
     "require": {
-        "php": ">=5.5.9",
-        "symfony/framework-bundle": "~2.7|~3.0",
-        "doctrine/annotations": "~1.1"
+        "php": "^7.1.3",
+        "doctrine/annotations": "~1.1",
+        "sensio/framework-extra-bundle": "^3.4||^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.0|^5.3",
-        "phpunit/php-code-coverage": "^2.2.0|^3.3.0|^4.0",
-        "symfony/phpunit-bridge": "~2.7|~3.0"
+        "phpunit/phpunit": "^6.4",
+        "phpunit/php-code-coverage": "^5.0",
+        "symfony/phpunit-bridge": "^3.4||^4.0",
+        "symfony/yaml": "^3.4||^4.0"
     },
     "extra": {
         "branch-alias": {
